+++ src/components/Routing.js
@@ -16,8 +16,8 @@ import { SetRecoveryInfoWithRouter } from './accounts/SetRecoveryInfo'
 import { RecoverAccountWithRouter } from './accounts/RecoverAccount'
 import { LoginWithRouter } from './login/Login'
 import { ContactsWithRouter } from './contacts/Contacts'
-import { AuthorizedAppsWithRouter } from './authorized-apps/AuthorizedApps'
-import { FullAccessKeysWithRouter } from './authorized-apps/AuthorizedApps'
+import { AuthorizedAppsWithRouter } from './access-keys/AccessKeys'
+import { FullAccessKeysWithRouter } from './access-keys/AccessKeys'
 import { SendMoneyWithRouter } from './send-money/SendMoney'
 import { ProfileWithRouter } from './profile/Profile'
 import { SignWithRouter } from './sign/Sign'
